- Addressed a number of warnings from the MSVC compiler. None of the changes should affect any behaviour; mostly addressing unreferenced parameters, unreachable code, constant expressions and potentially uninitialised variables. (note: until the next commit there may be some new warnings that appear, committing the fixes that are least likely to have side-effects first). - Fixed possible exception when importing WSC files containing item/character properties in an unexpected order. - Modified EF_FixCanSeeInClosedConts to always be active in debug builds, with additional logging to show situations where it takes effect.

+++ 0.56.b/common/CGrayData.cpp
@@ -157,7 +157,7 @@ CGrayItemInfo::CGrayItemInfo( ITEMID_TYPE id )
 	}
 
 	VERFILE_TYPE filedata;
-	long offset;
+	DWORD offset;
 	CUOIndexRec Index;
 	VERFILE_FORMAT format;
 	if ( g_VerData.FindVerDataBlock( VERFILE_TILEDATA, (id+TERRAIN_QTY)/UOTILE_BLOCK_QTY, Index ))
@@ -223,7 +223,7 @@ CGrayTerrainInfo::CGrayTerrainInfo( TERRAIN_TYPE id )
 	ASSERT( id < TERRAIN_QTY );
 
 	VERFILE_TYPE filedata;
-	long offset;
+	DWORD offset;
 	CUOIndexRec Index;
 	VERFILE_FORMAT format;
 	if ( g_VerData.FindVerDataBlock( VERFILE_TILEDATA, id/UOTILE_BLOCK_QTY, Index ))
